src/monitoring/drift_detector.py

from scipy.stats import ks_2samp
import numpy as np
from typing import Dict, Any, List

class DriftDetector:
    """Detects data and concept drift."""
    
    def __init__(self, reference_data: np.ndarray = None):
        self.reference_data = reference_data

    def detect_data_drift(self, current_data: np.ndarray, threshold: float = 0.05) -> Dict[str, Any]:
        """
        Performs Kolmogorov-Smirnov test to detect distribution drift.
        Assumes current_data is a feature matrix (e.g., TF-IDF vectors or reduced embeddings).
        For high-dimensional text data, monitoring specific aggregated features (like avg review length)
        or embedding clusters is often more practical than raw TF-IDF KS tests.
        """
        if self.reference_data is None:
             return {"drift_detected": False, "message": "No reference data."}

        # Simplified: Check drift on first component if available (e.g. SVD/PCA reduced) or just a placeholder
        # In a real scenario, we might track metadata statistics.
        
        # Example: Mock check
        return {"drift_detected": False, "p_value": 1.0}
