requirements.txt
tqdm
wordcloud
nltk
matplotlib
gensim
pandas
numpy
beautifulsoup4
Pillow
scikit-learn
python-box
ensure
PyYAML
mlflow
joblib
dvc
fastapi
uvicorn
pydantic
prometheus-fastapi-instrumentator
shap
